Move Service Worker code into utils folder
