+++ src/index.js
@@ -1,8 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+
 import App from './App';
-import registerServiceWorker from './registerServiceWorker';
+
+import registerServiceWorker from './utils/register-service-worker';
 
 ReactDOM.render(<App />, document.getElementById('root'));
-registerServiceWorker();
+
+// We may want to remove service workers for now as it may hinder our
+// productivity while developping the UI!
+//
+// registerServiceWorker();
