+++ src/utils/register-service-worker/index.js
@@ -1,0 +1,1 @@
+export { default, unregister } from './registerServiceWorker'
